consolidate trackingPointMethods, add API via implicits
