+++ dataflowengineoss/src/main/scala/io/shiftleft/dataflowengineoss/passes/reachingdef/ReachingDefProblem.scala
@@ -1,7 +1,8 @@
 package io.shiftleft.dataflowengineoss.passes.reachingdef
 
 import io.shiftleft.codepropertygraph.generated.nodes
-import io.shiftleft.semanticcpg.accesspath.{AccessPath, MatchResult, TrackedBase, TrackingPointMethods}
+import io.shiftleft.semanticcpg.accesspath.{AccessPath, MatchResult, TrackedBase}
+import io.shiftleft.semanticcpg.language.nodemethods.TrackingPointMethodsBase.ImplicitsAPI
 import io.shiftleft.semanticcpg.language._
 import org.slf4j.{Logger, LoggerFactory}
 
@@ -122,7 +123,7 @@ class ReachingDefTransferFunction(method: nodes.Method) extends TransferFunction
 
     val baseToCalls: Map[TrackedBase, List[(nodes.Call, AccessPath)]] = method.start.call.l
       .map { call =>
-        val (base, path) = TrackingPointMethods.toTrackedBaseAndAccessPath(call)
+        val (base, path) = call.trackedBaseAndAccessPath
         (base, (call, path))
       }
       .groupBy(_._1)
@@ -131,7 +132,7 @@ class ReachingDefTransferFunction(method: nodes.Method) extends TransferFunction
     def allOtherInstancesOf(node: nodes.StoredNode): Set[nodes.StoredNode] = {
       node match {
         case call: nodes.Call =>
-          val (base, accessPath) = TrackingPointMethods.toTrackedBaseAndAccessPath(call)
+          val (base, accessPath) = call.trackedBaseAndAccessPath
           baseToCalls
             .getOrElse(base, Nil)
             .collect {
