+++ semanticcpg/src/test/scala/io/shiftleft/semanticcpg/accesspath/TrackingPointToAccessPathTests.scala
@@ -5,6 +5,7 @@ import io.shiftleft.codepropertygraph.generated._
 import org.scalatest.matchers.should.Matchers._
 import org.scalatest.wordspec.AnyWordSpec
 import overflowdb._
+import io.shiftleft.semanticcpg.language.nodemethods.TrackingPointMethodsBase.toTrackedAccessPath
 
 class TrackingPointToAccessPathTests extends AnyWordSpec {
 
@@ -59,7 +60,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genID(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), C("a"))
+      toTrackedAccessPath(call) shouldBe E(C("b"), C("a"))
     }
   }
 
@@ -71,7 +72,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genID(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), C("a"))
+      toTrackedAccessPath(call) shouldBe E(C("b"), C("a"))
     }
   }
 
@@ -83,7 +84,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genLit(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), C("a"))
+      toTrackedAccessPath(call) shouldBe E(C("b"), C("a"))
     }
     "overtaint with others" in {
       val call =
@@ -92,7 +93,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genID(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), V)
+      toTrackedAccessPath(call) shouldBe E(C("b"), V)
     }
   }
 
@@ -104,7 +105,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genLit(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), C("a"))
+      toTrackedAccessPath(call) shouldBe E(C("b"), C("a"))
     }
     "overtaint with others" in {
       val call =
@@ -113,7 +114,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genID(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), V)
+      toTrackedAccessPath(call) shouldBe E(C("b"), V)
     }
 
   }
@@ -123,7 +124,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
       val call =
         genCALL(g, Operators.indirection, genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), I)
+      toTrackedAccessPath(call) shouldBe E(C("b"), I)
     }
   }
 
@@ -132,7 +133,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
       val call =
         genCALL(g, Operators.addressOf, genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), A)
+      toTrackedAccessPath(call) shouldBe E(C("b"), A)
     }
   }
   // new style
@@ -145,7 +146,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genLit(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), C("a"))
+      toTrackedAccessPath(call) shouldBe E(C("b"), C("a"))
     }
 
     "work with FieldIdentifier" in {
@@ -155,7 +156,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genFID(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), C("a"))
+      toTrackedAccessPath(call) shouldBe E(C("b"), C("a"))
     }
 
     "overtaint with others" in {
@@ -165,7 +166,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genID(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), V)
+      toTrackedAccessPath(call) shouldBe E(C("b"), V)
     }
   }
 
@@ -177,7 +178,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genLit(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), I, C("a"))
+      toTrackedAccessPath(call) shouldBe E(C("b"), I, C("a"))
     }
 
     "work with FieldIdentifier" in {
@@ -187,7 +188,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genFID(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), I, C("a"))
+      toTrackedAccessPath(call) shouldBe E(C("b"), I, C("a"))
     }
 
     "overtaint with others" in {
@@ -197,7 +198,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genID(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), I, V)
+      toTrackedAccessPath(call) shouldBe E(C("b"), I, V)
     }
   }
 
@@ -209,7 +210,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genLit(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), C("a"))
+      toTrackedAccessPath(call) shouldBe E(C("b"), C("a"))
     }
 
     "work with FieldIdentifier" in {
@@ -219,7 +220,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genFID(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), C("a"))
+      toTrackedAccessPath(call) shouldBe E(C("b"), C("a"))
     }
 
     "overtaint with others" in {
@@ -229,7 +230,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genID(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), V)
+      toTrackedAccessPath(call) shouldBe E(C("b"), V)
     }
   }
 
@@ -241,7 +242,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genLit(g, "12"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), S(12), I)
+      toTrackedAccessPath(call) shouldBe E(C("b"), S(12), I)
     }
 
     "work with FieldIdentifier" in {
@@ -251,7 +252,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genFID(g, "12"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), S(12), I)
+      toTrackedAccessPath(call) shouldBe E(C("b"), S(12), I)
     }
 
     "overtaint with others" in {
@@ -261,7 +262,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genID(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), VS, I)
+      toTrackedAccessPath(call) shouldBe E(C("b"), VS, I)
     }
     "overtaint on parsing failure" in {
       val call =
@@ -270,7 +271,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genLit(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), VS, I)
+      toTrackedAccessPath(call) shouldBe E(C("b"), VS, I)
     }
 
   }
@@ -284,7 +285,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genLit(g, "12"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), S(12))
+      toTrackedAccessPath(call) shouldBe E(C("b"), S(12))
     }
 
     "work with FieldIdentifier" in {
@@ -294,7 +295,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genFID(g, "12"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), S(12))
+      toTrackedAccessPath(call) shouldBe E(C("b"), S(12))
     }
 
     "overtaint with others" in {
@@ -304,7 +305,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genID(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), VS)
+      toTrackedAccessPath(call) shouldBe E(C("b"), VS)
     }
     "overtaint with parsing fails" in {
       val call =
@@ -313,7 +314,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genLit(g, "abc"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), VS)
+      toTrackedAccessPath(call) shouldBe E(C("b"), VS)
     }
 
   }
@@ -327,7 +328,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genLit(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), I, C("a"), A)
+      toTrackedAccessPath(call) shouldBe E(C("b"), I, C("a"), A)
     }
 
     "work with FieldIdentifier" in {
@@ -337,7 +338,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genFID(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), I, C("a"), A)
+      toTrackedAccessPath(call) shouldBe E(C("b"), I, C("a"), A)
     }
 
     "overtaint with others" in {
@@ -347,7 +348,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genID(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E(C("b"), I, V, A)
+      toTrackedAccessPath(call) shouldBe E(C("b"), I, V, A)
     }
   }
 
@@ -359,7 +360,7 @@ class TrackingPointToAccessPathTests extends AnyWordSpec {
                 genID(g, "a"),
                 genCALL(g, Operators.computedMemberAccess, genLit(g, "b"), genCALL(g, "foo")))
 
-      TrackingPointMethods.toTrackedAccessPath(call) shouldBe E()
+      toTrackedAccessPath(call) shouldBe E()
     }
   }
